HW2/requirements.txt
numpy
pandas
scipy
jupyterlab
matplotlib
seaborn
youtube-dl
tqdm
ffmpeg-python
soundfile
librosa
yt-dlp